requirements.txt
accelerate==1.7.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==13.1.0
blinker==1.9.0
Byaldi==0.0.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.0
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
-e git+https://github.com/zakkosi/ProfKim.git@2ae2147ef244b2e646a4cfa45af511e7fa916487#egg=colpali_engine&subdirectory=colpali
contourpy==1.3.1
cycler==0.12.1
datasets==3.6.0
dds-cloudapi-sdk==0.2.5
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
faiss==1.9.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.6.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
iopath==0.1.10
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_dtypes==0.5.1
mpmath==1.3.0
mteb==1.6.35
multidict==6.4.4
multiprocess==0.70.16
narwhals==1.41.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
ollama==0.4.9
omegaconf==2.3.0
openai==0.28.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
peft==0.15.2
pillow==10.2.0
platformdirs==4.3.6
portalocker==3.0.0
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytrec_eval-terrier==0.5.7
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
safetensors==0.4.5
SAM-2==1.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
srsly==2.5.1
streamlit==1.45.1
supervision==0.25.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.9.20
tiktoken==0.9.0
timm==1.0.11
tokenizers==0.15.2
toml==0.10.2
tomli==2.1.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.5.1
tqdm==4.67.0
transformers==4.38.2
transformers-stream-generator==0.0.5
triton==2.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.0
